+++ CHANGELOG.md
@@ -663,3 +663,4 @@
 * Define `oo_yukl` operator
 * Define `match` expression
 * Define `on` expression
+* Define `Valid` pattern synonym